Quantities

The qty function makes it easy to keep track of different quantities in chemical calculations. Metric prefixes are fully supported, i.e. any unit can be combined with standard metric scaling (mL, nmol, µM, etc.). Some quantities can also be used in common arithmetic operations.

qty(x, unit, scale_to_best_metric = TRUE)

get_qty_units(q)

Arguments

x
the numeric value of the quantity, can be a single value or a vector

unit
the unit of the quantity

scale_to_best_metric
whether to automatically scale to the best metric prefix

q
quantity or list of quantities

Details

amount: base unit mol but also understands mole, all metric prefixes allowed

mass: base unit g, all metric prefixes allowed

molecular weight: base unit g/mol, all metric prefixes allowed in the numerator

concentration (molarity): base unit M but also understands mol/L, all metric prefixes allowed in the numerator

concentration (density): base unit g/L but also understands g/l, all metric prefixes allowed in the numerator

volume: base unit L but also understands l, all metric prefixes allowed

pressure: base unit bar but also understands Pa, all metric prefixes allowed, the common non-metric units atm and psi, Torr and mTorr are also supported and will be automatically converted to bar

temperature: base unit K but also understands C and F and converts them to Kelvin

Functions

qty: generate a quantity object

get_qty_units: get units from a quantity or list of quantities (returns NA for objects that are not quantities)

See also

Examples

qty(0.045, "mmol/L")

#> Molarity [µM]
#> [1] 45

qty(6, "psi")

#> Pressure [mbar]
#> [1] 413.6855

qty(30, "C")

#> Temperature [K]
#> [1] 303.15